CI: drop support for clang-format10 and later
The only acceptable versions of clang-format are 8 and 9 for now
(until another round of big-churn-from-formatting, at which point
we'll update the required version).

clang-format-9 says:

    SpacesInSquareBrackets (bool)
      If true, spaces will be inserted after [ and before ]. Lambdas
      or unspecified size array declarations will not be affected.

clang-format-10 changes part of that to:

      Lambdas without arguments or unspecified size array
      declarations will not be affected.

This means that 9 will only allow `[name]` for captures, and 10
will only allow `[ name ]` for captures, so they ping-pong all
the lambda's in the codebase back and forth. Just don't.