During installation, if a swap partition exists, enable swap.
If there is no swap at all when the partition of the installation destination device is mounted, if a Swap partition exists, enable swap.
target_env_call cannot be used because it is before the files are copied to the installation destination.
host_env_process_output cannot be used because the return value cannot be obtained.
To avoid reinventing the wheel as much as possible, use subprocess.check_output.