Changes: note keyboard fixes
Neal Gompa suggested running KWin with the --locale1 flag, which
makes it listen to locale1 changes. For systems that don't support
that, there is now a different approach that rudely writes the
config file -- which you should probably then copy into the
target system later with a *preservefiles* module.

FIXES #2264