Should fix grub not adding lang section in /boot/grub/grub.cfg file as chroot take host env LANG variable and this value is usualy LANG=C.UTF-8 when booting a live installer.
Pb is only visible at boot when prompted to unlock Luks partition.
In this case, screen show a IE keymap instead of keymap of LANG selected
by user in Calamares.

Note that in this case, on ubuntu,  a "setupcon --save-only" and "update-initramfs -u" may be missing to
generate /etc/console-setup/cached_UTF-8_del.kmap.gz and make this file
integrated in initramfs to be able to unlock Luks partitions with the keymap selected in
calamares.